+++ src/aks-preview/azext_aks_preview/custom.py
@@ -1198,3 +1198,7 @@ def _update_addons(cmd, instance, subscription_id, resource_group_name, addons, 
     instance.aad_profile = None
 
     return instance
+
+
+def aks_get_versions(cmd, client, location):
+    return client.list_orchestrators(location, resource_type='managedClusters')