+++ src/aks-preview/azext_aks_preview/vendored_sdks/azure_mgmt_preview_aks/v2019_04_01/operations/__init__.py
@@ -12,9 +12,11 @@
 from .operations import Operations
 from .managed_clusters_operations import ManagedClustersOperations
 from .agent_pools_operations import AgentPoolsOperations
+from .container_services_operations import ContainerServicesOperations
 
 __all__ = [
     'Operations',
     'ManagedClustersOperations',
     'AgentPoolsOperations',
+    'ContainerServicesOperations',
 ]