+++ src/aks-preview/HISTORY.md
@@ -2,6 +2,12 @@
 
 Release History
 ===============
+0.4.1
++++++
+* Add `az aks get-versions -l location` to allow users to see all managed cluster versions.
+* Add `az aks get-upgrades` to get all available versions to upgrade.
+* Add '(preview)' suffix if kubernetes version is preview when using `get-versions` and `get-upgrades`
+
 0.4.0
 +++++
 * Add support for Azure policy add-on.